use bigint/int8 instead for compatability with storing chromaprints default unsigned ints

+++ acoustid--1.0.sql
@@ -1,16 +1,16 @@
 -- complain if script is sourced in psql, rather than via CREATE EXTENSION
 \echo Use "CREATE EXTENSION acoustid" to load this file. \quit
 
-CREATE OR REPLACE FUNCTION acoustid_compare2(int4[], int4[], int DEFAULT 0) RETURNS float4
+CREATE OR REPLACE FUNCTION acoustid_compare2(int8[], int8[], int DEFAULT 0) RETURNS float4
     AS 'MODULE_PATHNAME'
     LANGUAGE C VOLATILE STRICT -- marked as VOLATILE to avoid multiple calls, even though the results are immutable
     COST 1000;
 
-CREATE OR REPLACE FUNCTION acoustid_compare3(int4[], int4[], int DEFAULT -1) RETURNS float4
+CREATE OR REPLACE FUNCTION acoustid_compare3(int8[], int8[], int DEFAULT -1) RETURNS float4
     AS 'MODULE_PATHNAME'
     LANGUAGE C VOLATILE STRICT -- marked as VOLATILE to avoid multiple calls, even though the results are immutable
     COST 1000;
 
-CREATE OR REPLACE FUNCTION acoustid_extract_query(int4[]) RETURNS int4[]
+CREATE OR REPLACE FUNCTION acoustid_extract_query(int8[]) RETURNS int8[]
     AS 'MODULE_PATHNAME'
     LANGUAGE C IMMUTABLE STRICT;
